requirements-docker.txt
# Minimal BlockSafe Dependencies for Docker
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.22
google-genai==1.4.0
requests==2.32.4
python-dotenv==1.0.1
anyio==4.6.0
typing-extensions==4.12.2